Attempted smoothing of noise, and separated world generation ot its own file for further testing
