+++ src/js/utils.js
@@ -1,14 +1,13 @@
 // Diamond-Square Algorithm based from https://github.com/IceCreamYou/THREE.Terrain
-function generateDiamondSquare(map, segmentNumber, minHeight, maxHeight) {
+function generateNoise(map, segmentNumber, minHeight, maxHeight) {
     // Set the segment length to the smallest power of 2 that is greater than
     // the number of vertices in either dimension of the plane
-    // var segments = THREE.Math.ceilPowerOfTwo(segmentNumber + 1);
     var segments = Math.pow(2, Math.ceil(Math.log2(segmentNumber + 1)));
 
     // Initialize heightMap
     var size = segments + 1,
         heightMap = [],
-        smoothing = (maxHeight - minHeight),
+        smoothingRange = (maxHeight - minHeight), // Renamed from 'smoothing' to avoid confusion
         i,
         j,
         xl = segmentNumber + 1,
@@ -18,19 +17,20 @@ function generateDiamondSquare(map, segmentNumber, minHeight, maxHeight) {
     }
 
     // Generate heightMap
+    let currentSmoothing = smoothingRange;
     for (var l = segments; l >= 2; l /= 2) {
         var half = Math.round(l * 0.5),
             whole = Math.round(l),
             x,
             y,
             avg,
-            d,
-            e;
-        smoothing /= 2;
+            d;
+        // Removed 'e' as it was unused
+        currentSmoothing /= 2;
         // Square
         for (x = 0; x < segments; x += whole) {
             for (y = 0; y < segments; y += whole) {
-                d = Math.random() * smoothing * 2 - smoothing;
+                d = Math.random() * currentSmoothing * 2 - currentSmoothing;
                 avg = heightMap[x][y] +
                     heightMap[x + whole][y] +
                     heightMap[x][y + whole] +
@@ -42,7 +42,7 @@ function generateDiamondSquare(map, segmentNumber, minHeight, maxHeight) {
         // Diamond
         for (x = 0; x < segments; x += half) {
             for (y = (x + half) % l; y < segments; y += l) {
-                d = Math.random() * smoothing * 2 - smoothing;
+                d = Math.random() * currentSmoothing * 2 - currentSmoothing;
                 avg = heightMap[(x - half + size) % size][y] +
                     heightMap[(x + half) % size][y] +
                     heightMap[x][(y + half) % size] +
@@ -56,12 +56,56 @@ function generateDiamondSquare(map, segmentNumber, minHeight, maxHeight) {
         }
     }
 
+    // Smoothing pass
+    const smoothingIterations = 3;
+    for (let iter = 0; iter < smoothingIterations; iter++) {
+        const tempSmoothedHeightMap = [];
+        for (let r = 0; r <= segments; r++) {
+            // Initialize rows for the temporary smoothed map
+            tempSmoothedHeightMap[r] = new Float64Array(segments + 1);
+        }
+
+        // Assuming heightMap is [columnIndex][rowIndex]
+        for (let c = 0; c <= segments; c++) { // c for column (x-axis)
+            for (let r = 0; r <= segments; r++) { // r for row (y-axis)
+                let sumOfHeights = 0;
+                let numNeighbors = 0;
+
+                // Iterate over a 3x3 window (center + 8 neighbors)
+                for (let dc = -1; dc <= 1; dc++) { // delta column
+                    for (let dr = -1; dr <= 1; dr++) { // delta row
+                        const nc = c + dc; // neighbor column
+                        const nr = r + dr; // neighbor row
+
+                        // Check bounds
+                        if (nc >= 0 && nc <= segments && nr >= 0 && nr <= segments) {
+                            sumOfHeights += heightMap[nc][nr];
+                            numNeighbors++;
+                        }
+                    }
+                }
+                if (numNeighbors > 0) {
+                    tempSmoothedHeightMap[c][r] = sumOfHeights / numNeighbors;
+                } else {
+                    // This case should ideally not be hit if the center point is always included
+                    tempSmoothedHeightMap[c][r] = heightMap[c][r];
+                }
+            }
+        }
+        // Copy smoothed values back to heightMap for the next iteration or final use
+        for (let c = 0; c <= segments; c++) {
+            for (let r = 0; r <= segments; r++) {
+                heightMap[c][r] = tempSmoothedHeightMap[c][r];
+            }
+        }
+    }
+
     // Apply heightMap
-    for (i = 0; i < xl; i++) {
-        for (j = 0; j < yl; j++) {
+    for (i = 0; i < xl; i++) { // i is columnIndex
+        for (j = 0; j < yl; j++) { // j is rowIndex
             map[j * xl + i] += heightMap[i][j];
         }
     }
 }
 
-export { generateDiamondSquare };
+export { generateNoise };
